requirements.txt
openai>=1.0.0
google-auth>=2.0.0
google-api-python-client>=2.0.0
gspread>=5.0.0
requests>=2.28.0
python-dotenv>=1.0.0
tqdm>=4.64.0
Pillow>=9.0.0
wordpress-api>=1.2.9
python-wordpress-xmlrpc>=2.3
mysql-connector-python>=8.0.0
pandas>=2.0.0
